+++ CHANGELOG.md
@@ -3,13 +3,14 @@
 The changelog file describes changes made since v2.0.0, which made significant changes
 to the versions 1.x.x.
 
-### Unreleased 2.1.0
+### 2.1.0
 **Features:** new system for module/plugin building, improvements of annotation listing features,
 support for generic annotation visual style changes.
 
 **Maintenance:** removed outdated plugins.
 
 **Bugfixes:** plugins use also Cache API, annotation visuals updated also with history.
+Fix oidc login with events.
 
 ### 2.0.4
 **Features:** vertical magnification slider, allow 2x artificial zoom, annotation areas.